pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "larkmd"
version = "0.1.0.dev0"
description = "Stateful mirror from a git markdown tree to a Feishu/Lark wiki — incremental sync, transactional updates, cross-doc link patching, table column-width memory, mermaid preprocessing."
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "dzwalker", email = "david@zwalker.me"}]
requires-python = ">=3.10"
keywords = ["feishu", "lark", "wiki", "markdown", "sync", "docx"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
    "mypy",
    "build",
    "twine",
]

[project.scripts]
larkmd = "larkmd.cli:main"

[project.urls]
Homepage = "https://github.com/dzwalker/larkmd"
Issues = "https://github.com/dzwalker/larkmd/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/larkmd"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
